refactor(jedi): move S3 backup logic into jedi::cloud::s3 behind aws feature
axum-kbve becomes a thin HTTP/auth shell. The reusable S3 client + backup
summarize logic (from #14244/#14302) now lives in jedi::cloud::s3, gated by
a new 'aws' Cargo feature (dep:aws-config + dep:aws-sdk-s3) so only
axum-kbve pulls the AWS SDK — other jedi consumers are unaffected.

- jedi: new src/cloud/{mod,s3}.rs (S3Config, make_client, list_page,
  list_all, summarize + 5 unit tests). Feature aws.
- axum-kbve: delete s3backup/{client,summary}.rs; routes.rs imports from
  jedi::cloud::s3; drop direct aws-sdk-s3/aws-config deps (now via jedi
  feature). Staff gate + routes unchanged.

cargo test -p jedi --features aws (cloud) 5/5; cargo check -p axum-kbve OK.
