+++ apps/kbve/axum-kbve/src/s3backup/mod.rs
@@ -1,3 +1,1 @@
-pub mod client;
 pub mod routes;
-pub mod summary;
